It turns out the externalproject_add wasn't suitable for in process dependency.
It turns out the externalproject_add wasn't suitable for in process dependency.
The goal is to find dependency from provided path, then find it in dependency folder (if user already install), and if they are all empty, then it will install during build time.
But it turns out that ExternalProject_Add isn't for this, because it only installs during `--build`, should use FetchContent to do this.